+++ common/darwin/Classes/FlutterWebRTCPlugin.m
@@ -1,7 +1,9 @@
 #import "FlutterWebRTCPlugin.h"
 #import "AudioUtils.h"
 #import "CameraUtils.h"
+
 #import "FlutterRTCDataChannel.h"
+#import "FlutterDataPacketCryptor.h"
 #import "FlutterRTCDesktopCapturer.h"
 #import "FlutterRTCMediaStream.h"
 #import "FlutterRTCPeerConnection.h"
@@ -21,6 +23,9 @@
 #import <WebRTC/RTCFieldTrials.h>
 #import <WebRTC/WebRTC.h>
 #import <AVKit/AVKit.h>
+
+#import <WebRTC/RTCLogging.h>
+#import <WebRTC/RTCCallbackLogger.h>
 
 #import "LocalTrack.h"
 #import "LocalAudioTrack.h"
@@ -113,6 +118,8 @@ void postEvent(FlutterEventSink _Nonnull sink, id _Nullable event) {
 #if TARGET_OS_IPHONE
   FLutterRTCVideoPlatformViewFactory *_platformViewFactory;
 #endif
+
+  RTC_OBJC_TYPE(RTCCallbackLogger) * loggerCallback;
 }
 
 static FlutterWebRTCPlugin *sharedSingleton;
@@ -191,6 +198,7 @@ static FlutterWebRTCPlugin *sharedSingleton;
   self.localTracks = [NSMutableDictionary new];
   self.renders = [NSMutableDictionary new];
   self.frameCryptors = [NSMutableDictionary new];
+  self.dataCryptors = [NSMutableDictionary new];
   self.keyProviders = [NSMutableDictionary new];
   self.videoCapturerStopHandlers = [NSMutableDictionary new];
   self.recorders = [NSMutableDictionary new];
@@ -203,14 +211,10 @@ static FlutterWebRTCPlugin *sharedSingleton;
                                                name:AVAudioSessionRouteChangeNotification
                                              object:session];
 #endif
-#if TARGET_OS_OSX
-  [_peerConnectionFactory.audioDeviceModule setDevicesUpdatedHandler:^(void) {
-    NSLog(@"Handle Devices Updated!");
-    if (self.eventSink) {
-      postEvent( self.eventSink, @{@"event" : @"onDeviceChange"});
-    }
-  }];
-#endif
+
+  // Observe audio device module events.
+  _peerConnectionFactory.audioDeviceModule.observer = self;
+
   return self;
 }
 
@@ -267,9 +271,42 @@ static FlutterWebRTCPlugin *sharedSingleton;
 #endif
 }
 
+-(void) initLoggerCallback:(RTCLoggingSeverity)severity {
+  if(loggerCallback == nil) {
+    loggerCallback = [RTC_OBJC_TYPE(RTCCallbackLogger) new];
+    [loggerCallback start:^(NSString *logMessage) {
+      postEvent(self.eventSink, @{
+        @"event" : @"onLogData",
+        @"data" : logMessage
+      });
+    }];
+  }
+
+  loggerCallback.severity = severity;
+}
+
+-(RTCLoggingSeverity)str2LogSeverity:(NSString*)str {
+  if ([@"verbose" isEqualToString:str]) {
+    return RTCLoggingSeverityVerbose;
+  } else if ([@"info" isEqualToString:str]) {
+    return RTCLoggingSeverityInfo;
+  } else if ([@"warning" isEqualToString:str]) {
+    return RTCLoggingSeverityWarning;
+  } else if ([@"error" isEqualToString:str]) {
+    return RTCLoggingSeverityError;
+  } else if ([@"none" isEqualToString:str]) {
+    return RTCLoggingSeverityNone;
+  }
+
+  return RTCLoggingSeverityNone;
+}
+
 - (void)initialize:(NSArray*)networkIgnoreMask
-bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
-    // RTCSetMinDebugLogLevel(RTCLoggingSeverityVerbose);
+    bypassVoiceProcessing:(BOOL)bypassVoiceProcessing
+                 severity:(RTCLoggingSeverity)severity {
+    // RTCSetMinDebugLogLevel(severity);
+    [self initLoggerCallback:severity];
+
     if (!_peerConnectionFactory) {
         VideoDecoderFactory* decoderFactory = [[VideoDecoderFactory alloc] init];
         VideoEncoderFactory* encoderFactory = [[VideoEncoderFactory alloc] init];
@@ -278,7 +315,8 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
             [[VideoEncoderFactorySimulcast alloc] initWithPrimary:encoderFactory fallback:encoderFactory];
 
         _peerConnectionFactory =
-            [[RTCPeerConnectionFactory alloc] initWithBypassVoiceProcessing:bypassVoiceProcessing
+            [[RTCPeerConnectionFactory alloc] initWithAudioDeviceModuleType:RTCAudioDeviceModuleTypeAudioEngine
+                                                      bypassVoiceProcessing:bypassVoiceProcessing
                                                              encoderFactory:simulcastFactory
                                                              decoderFactory:decoderFactory
                                                       audioProcessingModule:_audioManager.audioProcessingModule];
@@ -321,7 +359,14 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
     if (options[@"networkIgnoreMask"] != nil) {
       networkIgnoreMask = ((NSArray*)options[@"networkIgnoreMask"]);
     }
-    [self initialize:networkIgnoreMask bypassVoiceProcessing:enableBypassVoiceProcessing];
+    RTCLoggingSeverity severity = RTCLoggingSeverityNone;
+    if (options[@"logSeverity"] != nil) {
+      NSString* severityStr = ((NSString*)options[@"logSeverity"]);
+      severity = [self str2LogSeverity:severityStr];
+    }
+
+    [self initialize:networkIgnoreMask bypassVoiceProcessing:enableBypassVoiceProcessing
+                     severity:severity];
     result(@"");
   } else if([@"setVideoEffects" isEqualToString:call.method]) {
     NSDictionary* argsMap = call.arguments;
@@ -615,7 +660,7 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
     NSString* dataChannelId = argsMap[@"dataChannelId"];
 
     [self dataChannelGetBufferedAmount:peerConnectionId dataChannelId:dataChannelId result:result];
-  } 
+  }
   else if ([@"dataChannelClose" isEqualToString:call.method]) {
     NSDictionary* argsMap = call.arguments;
     NSString* peerConnectionId = argsMap[@"peerConnectionId"];
@@ -1557,6 +1602,11 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
                 message:[NSString stringWithFormat:@"Error: peerConnection not found!"]
                 details:nil]);
     }
+  } else if ([@"setLogSeverity" isEqualToString:call.method]) {
+    NSDictionary* argsMap = call.arguments;
+    NSString* severityStr = argsMap[@"severity"];
+    RTCLoggingSeverity severity = [self str2LogSeverity:severityStr];
+    [self initLoggerCallback:severity];
 #if TARGET_OS_IOS
   } else if ([@"startRecordToFile" isEqualToString:call.method]){
 
@@ -1591,9 +1641,64 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
                                                 details:nil]);
                 }
 #endif
+    } else if ([@"startLocalRecording" isEqualToString:call.method]) {
+      RTCAudioDeviceModule* adm = _peerConnectionFactory.audioDeviceModule;
+      // Run on background queue
+      dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0), ^{
+        NSInteger admResult = [adm initAndStartRecording];
+
+        // Return to main queue
+        dispatch_async(dispatch_get_main_queue(), ^{
+          if (admResult == 0) {
+            result(nil);
+          } else {
+            result([FlutterError
+                errorWithCode:[NSString stringWithFormat:@"%@ failed", call.method]
+                      message:[NSString stringWithFormat:@"Error: adm api failed with code: %ld",
+                                                         (long)admResult]
+                      details:nil]);
+          }
+        });
+      });
+    } else if ([@"stopLocalRecording" isEqualToString:call.method]) {
+      RTCAudioDeviceModule* adm = _peerConnectionFactory.audioDeviceModule;
+      // Run on background queue
+      dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0), ^{
+        NSInteger admResult = [adm stopRecording];
+
+        // Return to main queue
+        dispatch_async(dispatch_get_main_queue(), ^{
+          if (admResult == 0) {
+            result(nil);
+          } else {
+            result([FlutterError
+                errorWithCode:[NSString stringWithFormat:@"%@ failed", call.method]
+                      message:[NSString stringWithFormat:@"Error: adm api failed with code: %ld",
+                                                         (long)admResult]
+                      details:nil]);
+          }
+        });
+      });
+    } else if ([@"isVoiceProcessingEnabled" isEqualToString:call.method]) {
+      RTCAudioDeviceModule* adm = _peerConnectionFactory.audioDeviceModule;
+      NSNumber* admResult = [NSNumber numberWithBool:adm.isVoiceProcessingEnabled];
+      result(admResult);
+    } else if ([@"isVoiceProcessingBypassed" isEqualToString:call.method]) {
+      RTCAudioDeviceModule* adm = _peerConnectionFactory.audioDeviceModule;
+      NSNumber* admResult = [NSNumber numberWithBool:adm.isVoiceProcessingBypassed];
+      result(admResult);
+    } else if ([@"setIsVoiceProcessingBypassed" isEqualToString:call.method]) {
+      RTCAudioDeviceModule* adm = _peerConnectionFactory.audioDeviceModule;
+      NSNumber* value = call.arguments[@"value"];
+      adm.voiceProcessingBypassed = value.boolValue;
+      result(nil);
     } else {
-    [self handleFrameCryptorMethodCall:call result:result];
-  }
+      if([self handleFrameCryptorMethodCall:call result:result]) {
+          return;
+      } else {
+          [self handleDataPacketCryptorMethodCall:call result:result];
+      }
+    }
 }
 
 - (void)dealloc {
@@ -1966,7 +2071,7 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
     NSNumber* maxIPv6Networks = json[@"maxIPv6Networks"];
      config.maxIPv6Networks = [maxIPv6Networks intValue];
   }
-    
+
   // === below is private api in webrtc ===
   if (json[@"tcpCandidatePolicy"] != nil &&
       [json[@"tcpCandidatePolicy"] isKindOfClass:[NSString class]]) {
@@ -2184,7 +2289,7 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
       @"kind" : codec.kind
     }];
   }
-    
+
   NSString *degradationPreference = @"balanced";
   if(parameters.degradationPreference != nil) {
     if ([parameters.degradationPreference intValue] == RTCDegradationPreferenceMaintainFramerate ) {
@@ -2400,7 +2505,7 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
   NSArray<RTCRtpEncodingParameters*>* currentEncodings = parameters.encodings;
   // new encodings
   NSArray* newEncodings = [newParameters objectForKey:@"encodings"];
-    
+
   NSString *degradationPreference = [newParameters objectForKey:@"degradationPreference"];
 
   if( degradationPreference != nil) {
@@ -2503,4 +2608,14 @@ bypassVoiceProcessing:(BOOL)bypassVoiceProcessing {
     }
     return nil;
 }
+
+#pragma mark - RTCAudioDeviceModuleDelegate methods
+
+- (void)audioDeviceModuleDidUpdateDevices:(RTCAudioDeviceModule *)audioDeviceModule {
+    NSLog(@"audioDeviceModule did update devices");
+    if (self.eventSink) {
+      postEvent( self.eventSink, @{@"event" : @"onDeviceChange"});
+    }
+}
+
 @end