+++ lib/src/native/factory_impl.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 import 'package:webrtc_interface/webrtc_interface.dart';
 
 import '../desktop_capturer.dart';
+import 'data_packet_cryptor_impl.dart';
 import 'android/audio_configuration.dart';
 import 'desktop_capturer_impl.dart';
 import 'frame_cryptor_impl.dart';
@@ -187,3 +188,6 @@ DesktopCapturer get desktopCapturer => DesktopCapturerNative.instance;
 MediaDevices get mediaDevices => MediaDeviceNative.instance;
 
 FrameCryptorFactory get frameCryptorFactory => FrameCryptorFactoryImpl.instance;
+
+DataPacketCryptorFactory get dataPacketCryptorFactory =>
+    DataPacketCryptorFactoryImpl.instance;