+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 
 # Changelog
 
+[1.1.0]
+* Synced flutter-webrtc v0.14.2
+	* [Doc] fix: typo in package description (#1895)
+	* [Android] fix: Video recording crashing and freezing on Android 14 Devices (#1886)
+	* [Android] fix: Add audio recording for Android Platform (#1884)
+	* [Dart] fix: Removed outdated code to avoid UI not being displayed in Windows release mode (#1890)
+	* [Apple] fix: Fix compile warnings (#1887)
+	* [Android] fix: Ensure both video and audio tracks are added before starting the muxer (#1879)
+	* [Apple/Android] feat: Add H265/HEVC support.
+	* [Mobile/Desktop] feat: Support write logs with Logger (logger package) (#1891)
+	* [Android] fix: Reduce Recording Stop Delay and Prevent Encoder OOM Crashes (Android) (#1912)
+	* [Native/Web] feat: small setVolume addition (#1904)
+	* [Web] feat: Add texture-based video rendering for web (#1911)
+	* [Android] fix: RECORDINGS - Add fallback resolutions for unsupported stream frame sizes on low-end Android devices (#1900)
+	* [Android] fix: Update proguard-rules.pro (#1902)
+	* [Android] upgrade compileSdk to 36 as standard for 16kb pages support (#1925)
+	* [Apple/Android] Local recording API for Darwin and Android (#1880)
+	* [Apple/Android] Data Packet Cryptor Support.
+
 [1.0.12] - 2025-09-30
 * [Android] Changed the configuration used when creating EGL rendering context to one that supports alpha channel to fix an issue with Impeller blending background with video texture on some devices.
 * [Android] Migrate from onSurfaceDestroyed to onSurfaceCleanup for SurfaceProducer.Callback.